+++ src/dovi/mod.rs
@@ -29,7 +29,7 @@ pub fn clear_start_code_emulation_prevention_3_byte(data: &[u8]) -> Vec<u8> {
         .collect::<Vec<u8>>()
 }
 
-pub fn add_start_code_emulation_prevention_3_byte(mut data: &mut Vec<u8>) {
+pub fn add_start_code_emulation_prevention_3_byte(data: &mut Vec<u8>) {
     for i in 0..data.len() {
         if i > 2 && i < data.len() - 2 && data[i - 2] == 0 && data[i - 1] == 0 && data[i] <= 3 {
             data.insert(i, 3);
